Move package-specific metadata code to each package
That way, we don't have to require packages from a path in node_modules, which will break with yarn workspaces.
Add initModel.js to simplify init'ing each server model.
