+++ client/src/Models/Ringdown.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { isValueEmpty } from '../utils';
 import { PatientFieldData, ValidationState } from './PatientFieldData';
 import * as metadata from '../../../shared/metadata';
-import convertToPropType from '../../../shared/convertToPropType';
+import convertToPropType from '../utils/convertToPropType';
 import DeliveryStatus from '../../../shared/constants/DeliveryStatus';
 
 // define the fields that must all have valid input to make the ringdown valid.  the second array item is an optional function to determine
